[HOPS-644] change inodeID from int to long

+++ src/main/java/io/hops/metadata/ndb/dalimpl/hdfs/InMemoryFileInodeClusterj.java
@@ -26,9 +26,9 @@ public class InMemoryFileInodeClusterj
 
     @PrimaryKey
     @Column(name = ID)
-    int getInodeId();
+    long getInodeId();
 
-    void setInodeId(int inodeId);
+    void setInodeId(long inodeId);
 
     @Column(name = DATA)
     byte[] getData();
@@ -62,7 +62,7 @@ public class InMemoryFileInodeClusterj
   }
 
   @Override
-  public FileInodeData get(int inodeId) throws StorageException {
+  public FileInodeData get(long inodeId) throws StorageException {
     final HopsSession session = connector.obtainSession();
     FileInodeDataDTO dataDto = session.find(FileInodeDataDTO.class, inodeId);
     if (dataDto != null) {
